Add documentation; add tests for response
