+++ lib/translation_io/api/version.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module TranslationIO
   module API
     VERSION = "0.1.0"
